factor the updates command out
