+++ crates/mdbook-goals/Cargo.toml
@@ -5,6 +5,7 @@ edition = "2021"
 
 [dependencies]
 anyhow = "1.0.94"
+clap = "4.5.23"
 mdbook = "0.4.43"
 regex = "1.11.1"
 rust-project-goals = { version = "0.1.0", path = "../rust-project-goals" }
