+++ crates/rust-project-goals-cli-llm/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "rust-project-goals-cli-llm"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+anyhow = "1.0.94"
+aws-config = "1.5.8"
+aws-sdk-bedrock = "1.57.0"
+aws-sdk-bedrockruntime = "1.55.0"
+chrono = "0.4.39"
+clap = { version = "4.5.23", features = ["derive"] }
+rust-project-goals = { version = "0.1.0", path = "../rust-project-goals" }
+rust-project-goals-llm = { version = "0.1.0", path = "../rust-project-goals-llm" }
+handlebars = { version = "6.2.0", features = ["dir_source"] }
+serde = "1.0.216"
+serde_json = "1.0.133"
+comrak = "0.31.0"
+progress_bar = "1.0.6"
+tokio = { version = "1.42.0", features = ["full"] }
