+++ crates/rust-project-goals-cli-llm/src/templates.rs
