+++ crates/rust-project-goals-cli/src/main.rs
@@ -1,28 +1,27 @@
-use anyhow::Context;
+use anyhow::{bail, Context};
+use clap::Parser;
 use regex::Regex;
+use rust_project_goals_llm::Updates;
 use std::path::PathBuf;
-use structopt::StructOpt;
 use walkdir::WalkDir;
 use rust_project_goals::gh::issue_id::Repository;
 
 mod rfc;
 mod generate_json;
 mod team_repo;
-mod templates;
-mod updates;
 
-#[derive(StructOpt, Debug)]
+#[derive(clap::Parser, Debug)]
 #[structopt(about = "Project goal preprocessor")]
 struct Opt {
-    #[structopt(subcommand)]
+    #[command(subcommand)]
     cmd: Command,
 
     /// Repository to use if applicable
-    #[structopt(long, default_value = "rust-lang/rust-project-goals")]
+    #[arg(long, default_value = "rust-lang/rust-project-goals")]
     repository: Repository,
 }
 
-#[derive(StructOpt, Debug)]
+#[derive(clap::Subcommand, Debug)]
 #[allow(dead_code)]
 enum Command {
     /// Print the comment required to initiate FCP
@@ -37,22 +36,22 @@ enum Command {
 
         /// Number of milliseconds to pause between github commands
         /// to avoid rate limiting
-        #[structopt(long, default_value = "500")]
+        #[arg(long, default_value = "500")]
         sleep: u64,
 
         /// Without this option, no action is taken.
-        #[structopt(long)]
+        #[arg(long)]
         commit: bool,
     },
 
     /// Generate the project-goal-owners team based on the owners found in `paths`.
     TeamRepo {
         /// Paths to the directories containing the goals (e.g., `src/2024h2`)
-        #[structopt(required = true, min_values = 1)]
+        #[arg(required = true /* , min_values = 1  */)]
         path: Vec<PathBuf>,
 
         /// Paths to the teams repository checkout
-        #[structopt(required = true, long = "team-repo")]
+        #[arg(required = true, long = "team-repo")]
         team_repo_path: PathBuf,
     },
 
@@ -67,49 +66,20 @@ enum Command {
 
         /// Path to write the json (e.g., `book/html/api/milestone.json`).
         /// If not provided, writes to stdout.
-        #[structopt(long)]
+        #[arg(long)]
         json_path: Option<PathBuf>,
     },
 
     /// Generate markdown with the list of updates for each tracking issue.
     /// Collects updates
     Updates {
-        /// Milestone for which we generate tracking issue data (e.g., `2024h2`).
-        milestone: String,
-
-        /// Quick mode does not use an LLM to generate a summary.
-        #[structopt(long)]
-        quick: bool,
-
-        /// Quick mode does not use an LLM to generate a summary.
-        #[structopt(long)]
-        vscode: bool,
-
-        /// If specified, write the output into the given file.
-        #[structopt(long)]
-        output_file: Option<PathBuf>,
-
-        /// Start date for comments.
-        /// If not given, defaults to 1 week before the start of this month.
-        start_date: Option<chrono::NaiveDate>,
-
-        /// End date for comments.
-        /// If not given, no end date.
-        end_date: Option<chrono::NaiveDate>,
-
-        /// Set a custom model id for the LLM.
-        #[structopt(long)]
-        model_id: Option<String>,
-
-        /// Set a custom region.
-        #[structopt(long)]
-        region: Option<String>,
+        #[command(flatten)]
+        updates: Updates
     },
 }
 
-#[tokio::main]
-async fn main() -> anyhow::Result<()> {
-    let opt = Opt::from_args();
+fn main() -> anyhow::Result<()> {
+    let opt: Opt = Opt::parse();
 
     match &opt.cmd {
         Command::FCP { path } => {
@@ -146,28 +116,21 @@ async fn main() -> anyhow::Result<()> {
         } => {
             generate_json::generate_json(&opt.repository, &milestone, json_path)?;
         }
-        Command::Updates {
-            milestone,
-            output_file,
-            start_date,
-            end_date,
-            quick,
-            vscode,
-            model_id,
-            region,
-        } => {
-            updates::updates(
-                &opt.repository,
-                milestone,
-                output_file.as_deref(),
-                start_date,
-                end_date,
-                *quick,
-                *vscode,
-                model_id.as_deref(),
-                region.as_deref(),
-            )
-            .await?;
+        Command::Updates {updates } => {
+            // The updates command is compiled separately so that we don't have to
+            // build all the LLM stuff if we are not using it.
+            let status = std::process::Command::new("cargo")
+            .arg("run")
+            .arg("-p")
+            .arg("rust-project-goals-cli-llm")
+            .arg("-q")
+            .arg("--")
+            .arg(&opt.repository.to_string())
+            .arg(&serde_json::to_string(updates).unwrap())
+            .status()?;
+            if !status.success() {
+                bail!("subcommand failed");
+            }
         }
     }
