+++ crates/rust-project-goals-llm/Cargo.toml
@@ -4,8 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-anyhow = "1.0.94"
-aws-config = "1.5.8"
-aws-sdk-bedrock = "1.57.0"
-aws-sdk-bedrockruntime = "1.55.0"
-rust-project-goals = { version = "0.1.0", path = "../rust-project-goals" }
+chrono = { version = "0.4.39", features = ["serde"] }
+clap = "4.5.23"
+serde = "1.0.216"
+structopt = "0.3.26"
